+++ packages/vue/src/lib.ts
@@ -10,6 +10,8 @@ export * as Result from "@effect-rx/rx/Result"
 
 const reportRuntimeError_ = reportError("Runtime")
 
+const filters = ["TypeError: failed to fetch", "AbortError"]
+
 const determineLevel = (cause: Cause<unknown>) => {
   const sq = Cause.squash(cause)
   if (!isHttpClientError(sq)) {
@@ -18,8 +20,9 @@ const determineLevel = (cause: Cause<unknown>) => {
   switch (sq._tag) {
     case "RequestError":
       return sq.reason === "Transport" ? LogLevel.Info : undefined
+      case "ResponseError":
+        return sq.reason === "Decode" && filters.some(_ => sq.cause?.toString().includes(_)) ? LogLevel.Info : undefined
   }
-  return undefined
 }
 
 export const reportRuntimeError = (cause: Cause<unknown>, extras?: Record<string, unknown>) =>