+++ packages/infra/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @effect-app/infra
 
+## 2.44.2
+
+### Patch Changes
+
+- 36996ed: housekeeping: update packages
+- Updated dependencies [36996ed]
+  - effect-app@2.33.1
+
+## 2.44.1
+
+### Patch Changes
+
+- eed6358: more local imports
+
 ## 2.44.0
 
 ### Minor Changes